feat(workflow-editor): add repeat menu description component and integrate with email step
